requirements.txt
scrapy==2.11.0
scrapy-user-agents==0.1.1
pandas==2.0.3
numpy==1.25.0
matplotlib==3.8.0
seaborn==0.12.2
boto3==1.28.0
google-cloud-storage==2.11.0
jupyter==1.0.0
imdbpy
python-dotenv
psycopg2-binary